Refactor to simplify configs
