--- framework/src/main/java/com/github/dbmdz/flusswerk/framework/messagebroker/ConnectionConfigImpl.java
@@ -1,63 +1,0 @@
-package com.github.dbmdz.flusswerk.framework.messagebroker;
-
-import com.rabbitmq.client.Address;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-class ConnectionConfigImpl implements ConnectionConfig {
-
-  private String username;
-
-  private String password;
-
-  private String virtualHost;
-
-  private final List<Address> addresses;
-
-  ConnectionConfigImpl() {
-    addresses = new ArrayList<>();
-    setPassword("guest");
-    setUsername("guest");
-    setVirtualHost("/");
-  }
-
-  @Override
-  public String getUsername() {
-    return username;
-  }
-
-  public final void setUsername(String username) {
-    this.username = username;
-  }
-
-  @Override
-  public String getPassword() {
-    return password;
-  }
-
-  public final void setPassword(String password) {
-    this.password = password;
-  }
-
-  @Override
-  public String getVirtualHost() {
-    return virtualHost;
-  }
-
-  public final void setVirtualHost(String virtualHost) {
-    this.virtualHost = virtualHost;
-  }
-
-  public void addAddress(String host, int port) {
-    addresses.add(new Address(host, port));
-  }
-
-  @Override
-  public List<Address> getAddresses() {
-    if (addresses.isEmpty()) {
-      return Collections.singletonList(new Address("localhost", 5672));
-    }
-    return addresses;
-  }
-}
