+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/messagebroker/MessageBrokerConfig.java
@@ -2,15 +2,56 @@ package com.github.dbmdz.flusswerk.framework.messagebroker;
 
 import com.fasterxml.jackson.databind.Module;
 import com.github.dbmdz.flusswerk.framework.model.Message;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-public interface MessageBrokerConfig {
+class MessageBrokerConfig {
 
-  int getDeadLetterWait();
+  private int deadLetterWait;
 
-  int getMaxRetries();
+  private int maxRetries;
 
-  Class<? extends Message> getMessageClass();
+  private final List<Module> jacksonModules;
 
-  List<Module> getJacksonModules();
+  private Class<? extends Message> messageClass;
+
+  public MessageBrokerConfig() {
+    jacksonModules = new ArrayList<>();
+    setMaxRetries(5);
+    setDeadLetterWait(30 * 1000);
+    setMessageClass(Message.class);
+  }
+
+  public int getDeadLetterWait() {
+    return deadLetterWait;
+  }
+
+  public final void setDeadLetterWait(int milliseconds) {
+    this.deadLetterWait = milliseconds;
+  }
+
+  public int getMaxRetries() {
+    return maxRetries;
+  }
+
+  public final void setMaxRetries(int maxRetries) {
+    this.maxRetries = maxRetries;
+  }
+
+  public Class<? extends Message> getMessageClass() {
+    return messageClass;
+  }
+
+  public final void setMessageClass(Class<? extends Message> messageClass) {
+    this.messageClass = messageClass;
+  }
+
+  public void addJacksonModule(Module jacksonModule) {
+    this.jacksonModules.add(jacksonModule);
+  }
+
+  public List<Module> getJacksonModules() {
+    return Collections.unmodifiableList(jacksonModules);
+  }
 }
